feat: replace api hook
- use useQuery
- add fetch util
- remove obsolete consts and types
