+++ src/constants.ts
@@ -1,11 +1,5 @@
 export const API_URL = 'https://rickandmortyapi.com/api/character';
 
-export const PROMISE_STATUS = {
-  PENDING: 'pending',
-  REJECTED: 'rejected',
-  FULFILLED: 'fulfilled',
-} as const;
-
 export const LOCALSTORAGE_SEARCH_QUERY_KEY = 'search-query';
 
 export const ROUTES = {
